Select last save file by default
When Load Game dialog is used for first time,
it has 'My Rome' as default selected file name.
Regardless if such save even exists.

It is much more convenient to have last existing savefile selected
instead by default.
(last as when numbers, date or such are used in save files, higher
numbers usually mean more recent, it could be better to have this
selected by actual modtime of files, but that is more tricky to do)